# Conflicts:
#	app/src/main/res/values-ru/strings.xml
#	app/src/main/res/values-uk/strings.xml
#	app/src/main/res/values/strings.xml
#	app/src/test/java/com/kinectpro/whattowear/utils/ConvertingUtilsTest.kt

+++ app/src/main/java/com/kinectpro/whattowear/data/model/response/WeatherData.kt
@@ -31,7 +31,7 @@ data class Data(
     val precipIntensity: Double?,
     val precipProbability: Double?,
     val temperature: Double?,
-    val apparentTemperature: Double?,
+    val apparentTemperature: Float?,
     val apparentTemperatureHigh: Float?,
     val apparentTemperatureLow: Float?,
     val apparentTemperatureHighTime: Long?,
